Revert "Fix/leave thread option"
